+++ core/apis.go
@@ -389,6 +389,10 @@ func (s *administrationImpl) AdmGetApplicationAccountDevices(appID string, orgID
 	return s.app.admGetApplicationAccountDevices(appID, orgID, accountID, l)
 }
 
+func (s *administrationImpl) AdmGetServiceRegs(appID *string, orgID *string, l *logs.Log) ([]model.ServiceReg, error) {
+	return s.app.admGetServiceRegs(appID, orgID, l)
+}
+
 ///
 
 //encryptionImpl