+++ driver/web/conversions_user.go
@@ -71,7 +71,7 @@ func partialAccountToDef(item model.Account, params map[string]interface{}) *Def
 	if params != nil {
 		paramsData = &params
 	}
-	return &Def.PartialAccount{Id: item.ID, AppId: item.AppOrg.Application.ID, OrgId: item.AppOrg.Organization.ID,
+	return &Def.PartialAccount{Id: &item.ID, AppId: item.AppOrg.Application.ID, OrgId: item.AppOrg.Organization.ID,
 		FirstName: item.Profile.FirstName, LastName: item.Profile.LastName, HasPermissions: item.HasPermissions,
 		Permissions: permissions, Roles: roles, Groups: groups, AuthTypes: authTypes,
 		DateCreated: dateCreated, DateUpdated: dateUpdated, Params: paramsData}
@@ -113,7 +113,7 @@ func accountRoleToDef(item model.AccountRole) Def.AppOrgRole {
 		dateUpdated = &formatted
 	}
 
-	return Def.AppOrgRole{Id: item.Role.ID, Name: item.Role.Name, Description: &item.Role.Description, System: &item.Role.System, DateCreated: &dateCreated, DateUpdated: dateUpdated, Permissions: &permissions}
+	return Def.AppOrgRole{Id: &item.Role.ID, Name: item.Role.Name, Description: &item.Role.Description, System: &item.Role.System, DateCreated: &dateCreated, DateUpdated: dateUpdated, Permissions: &permissions}
 }
 
 func accountRolesToDef(items []model.AccountRole) []Def.AppOrgRole {
@@ -137,7 +137,7 @@ func accountGroupToDef(item model.AccountGroup) Def.AppOrgGroup {
 		dateUpdated = &formatted
 	}
 
-	return Def.AppOrgGroup{Id: item.Group.ID, Name: item.Group.Name, Description: &item.Group.Description, System: &item.Group.System, DateCreated: &dateCreated, DateUpdated: dateUpdated, Permissions: &permissions, Roles: &roles}
+	return Def.AppOrgGroup{Id: &item.Group.ID, Name: item.Group.Name, Description: &item.Group.Description, System: &item.Group.System, DateCreated: &dateCreated, DateUpdated: dateUpdated, Permissions: &permissions, Roles: &roles}
 }
 
 func accountGroupsToDef(items []model.AccountGroup) []Def.AppOrgGroup {